Updated devices structure to have easier imports. Added more methods to pxd file
