+++ .gitignore
@@ -15,6 +15,8 @@
 *.log
 
 ### Above combination will ignore all files without extension, except MakeFiles ###
+
+studentapi.c
 *.pyc
 *.pyo
 __pycache__/
