+++ executor/Makefile
@@ -10,3 +10,10 @@ simple: run_simple_python.c
 
 studentapi: studentapi.pyx studentapi.pxd
 	python3.6 setup.py build_ext -i
+
+test_api: studentapi
+	python3.6 test_studentapi.py
+
+clean:
+	rm studentapi.c
+	rm studentapi.cpython-36m-x86_64-linux-gnu.so
