+++ executor/executor.h
@@ -7,12 +7,16 @@
 #include <stdint.h>                        //for standard int types
 #include <unistd.h>                        //for sleep
 #include <pthread.h>                       //for POSIX threads
-#include <signal.h> // Used to handle SIGTERM, SIGINT, SIGKILL
+#include <signal.h>                        // Used to handle SIGTERM, SIGINT, SIGKILL
 #include "../runtime_util.h"  //for runtime constants (TODO: consider removing relative pathname in include)
+#include "../shm_wrapper/shm_wrapper.h"    // Shared memory wrapper to get/send device data
 
 // Received a signal to stop the process
 // Clean up shared memory
 void sigintHandler(int sig_num);
+
+// Initialization method for executor
+void executor_init();
