+++ executor/setup.py
@@ -2,14 +2,16 @@ from setuptools import Extension, setup
 from Cython.Build import cythonize
 
 sourcefiles = [
-    "studentapi.pyx",
-    "../logger/logger.c"
+    "../logger/logger.c",
+    "../device_handler/devices.c",
+    "../shm_wrapper/shm_wrapper.c",
+    "studentapi.pyx"
 ]
 
 setup(
     name="Student API",
     ext_modules = cythonize([
-        Extension("studentapi", sources=sourcefiles)
-    ]),
+        Extension("studentapi", sources=sourcefiles, libraries=['rt'])
+    ], compiler_directives={'language_level' : "3"}),
     zip_safe=False,
 )
